+++ client/package.json
@@ -1,0 +1,73 @@
+{
+    "name": "learning-plan-web-app-client",
+    "version": "0.1.0",
+    "dependencies": {
+        "@testing-library/jest-dom": "^5.17.0",
+        "@testing-library/react": "^13.4.0",
+        "@testing-library/user-event": "^13.5.0",
+        "identity-obj-proxy": "^3.0.0",
+        "assert": "^2.1.0",
+        "babel-loader": "^9.1.3",
+        "buffer": "^6.0.3",
+        "crypto-browserify": "^3.12.0",
+        "css-loader": "^6.8.1",
+        "file-loader": "^6.2.0",
+        "https-browserify": "^1.0.0",
+        "jsx-loader": "^0.13.2",
+        "path-browserify": "^1.0.1",
+        "react": "^18.2.0",
+        "react-app-rewired": "^2.2.1",
+        "react-dom": "^18.2.0",
+        "react-icons": "^4.11.0",
+        "react-router-dom": "^6.17.0",
+        "react-scripts": "^5.0.1",
+        "stream-browserify": "^3.0.0",
+        "stream-http": "^3.2.0",
+        "style-loader": "^3.3.3",
+        "tls": "^0.0.1",
+        "url": "^0.11.3",
+        "util": "^0.12.5",
+        "web-vitals": "^2.1.4"
+    },
+    "scripts": {
+        "start": "react-app-rewired start",
+        "build": "react-app-rewired build",
+        "test": "react-app-rewired test",
+        "eject": "react-app-rewired eject",
+        "dev": "webpack serve"
+    },
+    "eslintConfig": {
+        "extends": [
+          "react-app",
+          "react-app/jest"
+        ]
+    },
+    "browserslist": {
+    "production": [
+        ">0.2%",
+        "not dead",
+        "not op_mini all"
+    ],
+    "development": [
+        "last 1 chrome version",
+        "last 1 firefox version",
+        "last 1 safari version"
+    ]
+    },
+    "devDependencies": {
+        "@babel/core": "^7.22.9",
+        "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
+        "@babel/preset-env": "^7.23.2",
+        "@babel/preset-react": "^7.22.5",
+        "babel-preset-react": "^6.24.1",
+        "html-webpack-plugin": "^5.5.3",
+        "webpack": "^5.89.0",
+        "webpack-cli": "^5.1.4",
+        "webpack-dev-server": "^4.15.1"
+    },
+    "jest": {
+        "moduleNameMapper": {
+            "\\.(css|scss)$": "identity-obj-proxy"
+        }
+    }
+}
