+++ client/src/screens/login/index.js
@@ -1,0 +1,1 @@
+export {default} from "./login.jsx";
