+++ config-overrides.js
@@ -5,9 +5,9 @@ module.exports = function override(config) {
     Object.assign(fallback, {
         "stream": require.resolve("stream-browserify"),
         "path": require.resolve("path-browserify"),
-        "crypto": require.resolve("crypto-browserify"),
+        "crypto": false,
         "assert": require.resolve("assert"),
-        "os": require.resolve("os-browserify"),
+        "os": false,
         "http": require.resolve("stream-http"),
         "https": require.resolve("https-browserify"),
         "url": require.resolve("url"),
