+++ server/package.json
@@ -1,0 +1,31 @@
+{
+    "name": "learning-plan-web-app-server",
+    "version": "0.1.0",
+    "private": true,
+    "dependencies": {
+        "dotenv": "^16.3.1",
+        "jest": "^29.5.0",
+        "pg": "^8.11.3"
+    },
+    "scripts": {
+        "test": "jest --testPathPattern=tests --detectOpenHandles --verbose --coverage"
+    },
+    "eslintConfig": {
+        "extends": [
+          "react-app",
+          "react-app/jest"
+        ]
+    },
+    "browserslist": {
+        "production": [
+            ">0.2%",
+            "not dead",
+            "not op_mini all"
+        ],
+        "development": [
+            "last 1 chrome version",
+            "last 1 firefox version",
+            "last 1 safari version"
+        ]
+    }
+}
