+++ server/parser/databaseParser.js
@@ -25,7 +25,7 @@ export class DatabaseParser {
         const query = {
             text: "INSERT INTO ACCOUNT(username, account_password, email) VALUES($1, $2, $3)",
             values: [username, password, email]
-        }
+        };
         const client = await this.pool.connect();
         await client.query(query);
         client.release();
@@ -34,11 +34,12 @@ export class DatabaseParser {
     
     async retrieveLogin(username, password) {
         console.log("Retrieving login...");
+        // TODO: "TypeError: CloudflareSocket is not a constructor"
         const client = await this.pool.connect();
         const query = {
             text: "SELECT * FROM ACCOUNT WHERE username = $1 AND account_password = $2",
             values: [username, password]
-        }
+        };
         const result = await client.query(query);
         client.release();
         console.log("Login found!");
@@ -51,7 +52,7 @@ export class DatabaseParser {
         const query = {
             text: "SELECT * FROM PROFILE WHERE account_id = $1",
             values: [id]
-        }
+        };
         await client.query(query);
         const result = await client.query(`SELECT * FROM PROFILE WHERE account_id = ${id}`);
         client.release();
@@ -65,7 +66,7 @@ export class DatabaseParser {
         const query = {
             text: "INSERT INTO PROFILE(firstName, lastName, account_id) VALUES($1, $2, $3)",
             values: [firstName, lastName, account_id]
-        }
+        };
         await client.query(query);
         client.release();
         console.log("Profile created!");
@@ -77,7 +78,7 @@ export class DatabaseParser {
         const query = {
             text: "INSERT INTO PROFILE(firstName, lastName, profilePicture, jobTitle, bio) VALUES($1, $2, $3, $4, $5) WHERE account_id = $6",
             values: [firstName, lastName, profilePicture, jobTitle, bio, account_id]
-        }
+        };
         await client.query(query);
         client.release();
         console.log("Profile data saved!");
