+++ server/tests/controller/login.test.js
@@ -1,11 +1,11 @@
-import { DatabaseParser } from "../../src/parser/DatabaseParser";
-import { LoginAPI } from "../../src/controller/loginProcessor";
+import { DatabaseParser } from "../../client/src/parser/DatabaseParser";
+import { LoginAPI } from "../../client/src/controller/loginProcessor";
 
-jest.mock("../../src/parser/DatabaseParser", () => {
+jest.mock("../parser/DatabaseParser", () => {
     const testParser = {
         retrieveLogin: jest.fn()
     };
-    return { DatabaseParser: jest.fn(() => testParser)}
+    return { DatabaseParser: jest.fn(() => testParser) };
 });
 
 describe('Login Functions', () => {
@@ -27,7 +27,7 @@ describe('Login Functions', () => {
             username: "Xx_george_xX",
             password: "09122001",
             email: "George123@Gmail.com"
-        }
+        };
         parser.retrieveLogin.mockResolvedValueOnce(
             Promise.resolve([{account_id: testData.account_id, username: testData.username, password: testData.password, email: testData.email}])
         );
